+++ sites/partners/src/listings/PaperListingForm/UnitsSummaryForm.tsx
@@ -337,20 +337,6 @@ const UnitsSummaryForm = ({
           {rentType === "fixed" && (
             <>
               <GridCell>
-                <ViewItem label={t("t.minimumIncome")}>
-                  <Field
-                    id="minimumIncomeMin"
-                    name="minimumIncomeMin"
-                    label={t("t.minimumIncome")}
-                    placeholder="0.00"
-                    register={register}
-                    type="number"
-                    prepend="$"
-                    readerOnly
-                  />
-                </ViewItem>
-              </GridCell>
-              <GridCell>
                 <ViewItem label={t("listings.unitsSummary.monthlyRentMin")}>
                   <Field
                     id="monthlyRentMin"
@@ -366,6 +352,8 @@ const UnitsSummaryForm = ({
                     validation={{ required: true }}
                   />
                 </ViewItem>
+              </GridCell>
+              <GridCell>
                 <ViewItem label={t("listings.unitsSummary.monthlyRentMax")}>
                   <Field
                     id="monthlyRentMax"