+++ sites/public/pages/listings.tsx
@@ -66,8 +66,8 @@ const getListingTableData = (listing: Listing) => {
 const getListings = (listings: Listing[]) => {
   const unitSummariesHeaders = {
     unitType: t("t.unitType"),
-    minimumIncome: t("t.minimumIncome"),
     rent: t("t.rent"),
+    availability: t("t.availability"),
   }
   return listings.map((listing: Listing, index) => {
     return (