+++ ui-components/src/helpers/tableSummaries.tsx
@@ -3,43 +3,14 @@ import { t } from "./translator"
 import { UnitSummary, UnitsSummary } from "@bloom-housing/backend-core/types"
 import { GroupedTableGroup } from "../tables/GroupedTable"
 
-export const getSummaryRow = (
+const getSummaryRow = (
   totalAvailable: number,
-  minIncomeRangeMin?: string,
-  minIncomeRangeMax?: string,
   rentRangeMin?: string,
   rentRangeMax?: string,
   rentAsPercentIncomeRangeMin?: string,
   rentAsPercentIncomeRangeMax?: string,
   unitTypeName?: string
 ) => {
-  let minIncome = <></>
-  if (minIncomeRangeMin == undefined && minIncomeRangeMax == undefined) {
-    //TODO(#345): figure out what to display when there's no data
-    minIncome = <strong>{t("t.call")}</strong>
-  } else if (minIncomeRangeMin == minIncomeRangeMax || minIncomeRangeMax == undefined) {
-    minIncome = (
-      <>
-        <strong>{minIncomeRangeMin}</strong>
-        {t("t.perMonth")}
-      </>
-    )
-  } else if (minIncomeRangeMin == undefined) {
-    minIncome = (
-      <>
-        <strong>{minIncomeRangeMax}</strong>
-        {t("t.perMonth")}
-      </>
-    )
-  } else {
-    minIncome = (
-      <>
-        <strong>{minIncomeRangeMin}</strong> {t("t.to")} <strong>{minIncomeRangeMax}</strong>
-        {t("t.perMonth")}
-      </>
-    )
-  }
-
   const getRent = (rentMin?: string, rentMax?: string, percent = false) => {
     const unit = percent ? `% ${t("t.income")}` : ` ${t("t.perMonth")}`
     if (rentMin == undefined && rentMax == undefined) {
@@ -76,7 +47,6 @@ export const getSummaryRow = (
 
   return {
     unitType: <strong>{t(`listings.unitTypes.${unitTypeName}`)}</strong>,
-    minimumIncome: <>{minIncome}</>,
     rent: <>{rent}</>,
     availability: (
       <>
@@ -99,8 +69,6 @@ export const getSummariesTableFromUnitSummary = (summaries: UnitSummary[]) => {
     const unitSummaries = summaries.map((unitSummary) => {
       return getSummaryRow(
         unitSummary.totalAvailable ? unitSummary.totalAvailable : 0,
-        unitSummary.minIncomeRange.min,
-        unitSummary.minIncomeRange.max,
         unitSummary.rentRange.min,
         unitSummary.rentRange.max,
         unitSummary.rentAsPercentIncomeRange.min
@@ -128,8 +96,6 @@ export const getSummariesTableFromUnitsSummary = (summaries: UnitsSummary[]) => 
     const unitSummaries = summaries.map((unitSummary) => {
       return getSummaryRow(
         unitSummary.totalAvailable ? unitSummary.totalAvailable : 0,
-        unitSummary.minimumIncomeMin,
-        unitSummary.minimumIncomeMax,
         unitSummary.monthlyRentMin?.toString(),
         unitSummary.monthlyRentMax?.toString(),
         unitSummary.monthlyRentAsPercentOfIncome,